% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gPlot.R
\name{gPlot}
\alias{gPlot}
\title{gPlot}
\usage{
gPlot(fit_obj, spike_param)
}
\arguments{
\item{fit_obj}{Fitted object from function 'lsum'.}

\item{spike_param}{The value of the spike parameter to be specified, which must be in
\code{spike_params} designated in 'lsum' function. Each \code{spike_param} corresponds
to a unique gPlot. If default \code{spike_params} is used in function 'lsum', one can
check the value of \code{spike_params} by calling \verb{$spike_params} from the returned
object from function 'lsum'.}
}
\description{
Plot maximized \eqn{g(\bold{\beta},\theta,\sigma|\bold{\beta^{(t)}},\theta^{(t)},
\sigma^{(t)})} over iterations at a specified spike parameter.
The function \eqn{g} is a concave lower bound for objective log-likelihood function at
\eqn{\bold{\beta^{(t)}}}, \eqn{\theta^{(t)}} and \eqn{\sigma^{(t)}} by Jensen's Inequality.
\eqn{\bold{\beta^{(t+1)}}}, \eqn{\theta^{(t+1)}} and \eqn{\sigma^{(t+1)}}
are obtained by maximizing \eqn{g(\bold{\beta},\theta,\sigma|\bold{\beta^{(t)}},\theta^{(t)},
\sigma^{(t)})}. See Vignette for more details.
}
