% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathPlot.R
\name{pathPlot}
\alias{pathPlot}
\title{pathPlot}
\usage{
pathPlot(fit_obj, path_of = "beta")
}
\arguments{
\item{fit_obj}{The fitted object from 'lsum' function.}

\item{path_of}{Character. \code{path_of} = \code{"beta"}, \code{"theta"} or \code{"sigma"}.
Default is \code{"beta"}.}
}
\description{
Plot the path of \code{beta}, \code{theta} or \code{sigma} over \code{spike_params}.
}
